pyproject.toml
[project]
name = "hedera_sdk_python"
dynamic = ["version"]
description = "A Hiero SDK in pure Python"
maintainers = [
    { name = "Nadine Loepfe", email = "nadine.loepfe@hashgraph.com" }
]
authors = [
    { name = "Nadine Loepfe", email = "nadine.loepfe@hashgraph.com" },
    { name = "Richard Bair", email = "rbair23@users.noreply.github.com" }
]
license = {  file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.9.2"
dependencies = [
    "protobuf==5.27.2",
    "grpcio-tools==1.68.1",
    "grpcio==1.68.1",
    "cryptography==44.0.0",
    "python-dotenv==1.0.1",
    "requests==2.32.3"
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.9",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
lint = [
    "ruff>=0.8.3",
]

[tool.pdm]
version = {source = "scm"}

[tool.pdm.build]
package-dir = "src"
excludes = ["**/.pytest_cache"]

[tool.ruff]
line-length = 120

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
